+++ robotframework-ls/docs/changelog.md
@@ -6,6 +6,7 @@ NEXT
 - Recursively resolve variables and provide better error messages on failure conditions. [#715](https://github.com/robocorp/robotframework-lsp/issues/715)
 - If a resource/library/variable import is not found and its name has variables, show the resolved name.
 - Report error when multiple keywords with the same name are found. [#432](https://github.com/robocorp/robotframework-lsp/issues/432)
+  - Can be disabled with: `robot.lint.keywordResolvesToMultipleKeywords: false`
 - Fix issues in file system notifications. [#710](https://github.com/robocorp/robotframework-lsp/issues/710)
 - Fixed issue where doc highlight could raise error with unclosed variable.
 - Interactive console